Add `buildDatapack` action, refactor IDE actions, and integrate shortcut handling
