+++ src/core/compiler/index.ts
@@ -3,3 +3,4 @@ export {compileEditorState, inferCompilerType} from './emitter'
 export * from './ir'
 export * from './generator'
 export {orchestrate} from './orchestrator.ts'
+export type {OutputFiles} from './outputFiles.ts'
