+++ src/ui/ide/actions.tsx
@@ -7,6 +7,9 @@ import {loadProject, newProject, saveProject} from '../../core/save'
 import type {IDEContextValue} from './context/IDEContext.tsx'
 import {controller} from '../editor'
 import {TextDialogue} from './WorkspaceDialogues'
+import {orchestrate, type OutputFiles} from '../../core/compiler'
+import {mapToOutputZip} from '../../core/output-preview'
+import {useProjectConfigStore} from '../../stores'
 
 export const actions = {
   saveProject({ blocklyWorkspaceRef, setHasUnsavedFileChanges }: IDEContextValue) {
@@ -14,7 +17,7 @@ export const actions = {
     saveProject({workspace: blocklyWorkspaceRef.current})
     setHasUnsavedFileChanges(false)
   },
-  openProject({ blocklyWorkspaceRef, setHasUnsavedFileChanges }: IDEContextValue) {
+  openProject({ blocklyWorkspaceRef, setHasUnsavedFileChanges, setCompiledOutput, setOutputViewerOpen }: IDEContextValue) {
     controller.openEditorModal({
       title: 'Unsaved changes',
       editor: (
@@ -26,13 +29,17 @@ export const actions = {
       mode: 'confirm',
       onConfirm: () => {
         if (!blocklyWorkspaceRef.current) return
+
         loadProject({workspace: blocklyWorkspaceRef.current})
         setHasUnsavedFileChanges(false)
+
+        setCompiledOutput(null)
+        setOutputViewerOpen(false)
       },
       noFullscreen: true
     })
   },
-  newProject({ blocklyWorkspaceRef, setHasUnsavedFileChanges }: IDEContextValue) {
+  newProject({ blocklyWorkspaceRef, setHasUnsavedFileChanges, setCompiledOutput, setOutputViewerOpen }: IDEContextValue) {
     controller.openEditorModal({
       title: 'Unsaved changes',
       editor: (
@@ -44,10 +51,33 @@ export const actions = {
       mode: 'confirm',
       onConfirm: () => {
         if (!blocklyWorkspaceRef.current) return
+
         newProject(blocklyWorkspaceRef.current)
         setHasUnsavedFileChanges(true) // A new project is not yet saved to computer.
+
+        setCompiledOutput(null)
+        setOutputViewerOpen(false)
       },
       noFullscreen: true
     })
+  },
+  buildDatapack({ blocklyWorkspaceRef, setCompiledOutput, setOutputViewerOpen }: IDEContextValue) {
+    if (!blocklyWorkspaceRef.current) return
+
+    const outputFiles = orchestrate(
+      blocklyWorkspaceRef.current, useProjectConfigStore.getState().projectConfig)
+
+    // Replace namespace with readable name later
+    outputFiles.download(useProjectConfigStore.getState().projectConfig.namespace)
+
+    viewOutputFiles(outputFiles, { setCompiledOutput, setOutputViewerOpen })
   }
+}
+
+function viewOutputFiles(
+  outputFiles: OutputFiles,
+  { setCompiledOutput, setOutputViewerOpen }: Pick<IDEContextValue, 'setCompiledOutput' | 'setOutputViewerOpen'>
+) {
+  setCompiledOutput(mapToOutputZip(outputFiles.toStringMap(), new Date()))
+  setOutputViewerOpen(true)
 }
